changed namespace of pagiantion service, refactoring
