+++ src/backend/app/Http/Controllers/Api/v1/OrdersController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Api\v1\Order\IndexOrder;
 use App\Http\Requests\Api\v1\Order\StoreOrder;
 use App\Http\Requests\Api\v1\Order\UpdateOrder;
 use App\Models\Order;
+use App\Services\Api\v1\PaginationService;
 use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 
@@ -21,44 +22,24 @@ class OrdersController extends Controller
      * @throws \Exception
      */
     public function index(IndexOrder $request): JsonResponse {
-        $orderDirection = 'asc';
-        $perPage = 10;
-        $page = 1;
-        $orderBy = 'orders.id';
-
-        if($request->has('orderDirection')){
-            $orderDirection = $request->orderDirection;
-        }
-
-        if($request->has('perPage')) {
-            $perPage = $request->perPage;
-        }
-
-        if($request->has('page')){
-            $page = $request->page;
-        }
-
-        if($request->has('orderBy')){
-            $orderBy = $request->orderBy;
-        }
-
+        $paginationData = app(PaginationService::class)->getPagination($request);
         $from = new Carbon($request->input('from', Carbon::now()->subWeek()));
         $to = new Carbon($request->input('to', Carbon::now()));
 
         $orders = Order::whereBetween('start_time', [$from, $to])
             ->join('clients', 'clients.id', '=', 'orders.client_id')
             ->join('machines_and_procedures', 'orders.machine_id', '=', 'machines_and_procedures.id')
-            ->orderBy($orderBy, $orderDirection)
-            ->offset(($perPage * $page) - $perPage)
-            ->limit($perPage)
+            ->orderBy('orders.id', $paginationData['orderDirection'])
+            ->offset($paginationData['offset'])
+            ->limit($paginationData['perPage'])
             ->get();
 
         $data = [
             'items' => $orders,
             'total' => $orders->count(),
-            'perPage' => $perPage,
-            'currentPage' => $page,
-            'lastPage' => $orders->count() / $perPage,
+            'perPage' => $paginationData['perPage'],
+            'currentPage' => $paginationData['page'],
+            'lastPage' => $orders->count() / $paginationData['perPage'],
         ];
 
         return response()->json($data, 200);
