+++ src/backend/app/Services/Api/v1/PaginationService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Api\v1;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
